+++ HISTORY.rst
@@ -6,6 +6,12 @@ v0.32 (unreleased)
 ..................
 * support ``ForwardRef`` (without self-referencing annotations) in Python3.6, #706 by @koxudaxi
 
+v0.32 (unreleased)
+..................
+* add model name to ``ValidationError`` error message, #676 by @dmontagu
+* **breaking change**: remove ``__getattr__`` and rename ``__values__`` to ``__dict__`` on ``BaseModel``,
+  deprecation warning on use ``__values__`` attr, attributes access speed increased up to 14 times, #712 by @MrMrRobat
+
 v0.31.1 (2019-07-31)
 ....................
 * fix json generation for ``EnumError``, #697 by @dmontagu
@@ -23,6 +29,7 @@ v0.31 (2019-07-24)
 * add ``Config.keep_untouched`` for custom descriptors support, #679 by @MrMrRobat
 * use ``inspect.cleandoc`` internally to get model description, #657 by @tiangolo
 * add ``Color`` to schema generation, by @euri10
+* add documentation for Literal type, #651 by @dmontagu
 
 v0.30.1 (2019-07-15)
 ....................